Update Axum example
