+++ examples/axum_example/api/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "axum-example-api"
+version = "0.1.0"
+authors = ["Yoshiera Huang <huangjasper@126.com>"]
+edition = "2021"
+publish = false
+
+[dependencies]
+axum-example-core = { path = "../core" }
+tokio = { version = "1.18.1", features = ["full"] }
+axum = "0.5.4"
+tower = "0.4.12"
+tower-http = { version = "0.3.3", features = ["fs"] }
+tower-cookies = "0.6.0"
+anyhow = "1.0.57"
+dotenv = "0.15.0"
+serde = "1.0.137"
+serde_json = "1.0.81"
+tera = "1.15.0"
+tracing-subscriber = { version = "0.3.11", features = ["env-filter"] }
+entity = { path = "../entity" }
+migration = { path = "../migration" }
