+++ examples/axum_example/api/src/flash.rs
