+++ examples/axum_example/api/static/images/favicon.png
