+++ examples/axum_example/core/src/lib.rs
@@ -1,0 +1,7 @@
+mod mutation;
+mod query;
+
+pub use mutation::*;
+pub use query::*;
+
+pub use sea_orm;
